Basic Vuex data flow now working
This incorporates a very simple Vuex data-flow. Here's what's going on:

Currently, this app requires one bit of shared stateful data: the collection of
books that is available at any given time. This data is loaded asynchronously
from an external source (the Standard Ebooks opds file, which is parsed into
JSON after being read).

So we need to have multiple components pointing to the same source of data (we
only want to request it once), and they need to be notified when that data
changes (otherwise, they initially are pointing to an empty array and then never
update their views with the data once it is available).

The Vuex library makes this pretty simple to do. A "store" object is defined in
the main.js file (along with the router), and like the router object it is
passed to the new Vue instance upon initialization. This makes the store object
and its properties available in all child components, as $store.state.

We don't want anything changing this object directly. Instead
"mutations" (synchronous operations which directly change the state object) and
"actions" (asynchronous operations which can be dispatched like events, and
which commit mutations) are used. Actions call mutations; all other components
should only use actions (store.dispatch("action", otherArgs) is how it's done).

To read the always-up-to-date data in the store, access the part you need as a
computed property in a component. This means that whenever the store changes,
the component watches and re-renders whatever is needed.

That's it - welcome to the Flux architecture!
