+++ package.json
@@ -108,6 +108,7 @@
   },
   "dependencies": {
     "lodash": "^4.17.4",
+    "vuex": "^2.3.0",
     "xml2js": "^0.4.17"
   }
 }
